Flatten namespace [skip ci]
