+++ src/ImageProcessorCore/ImageExtensions.cs
@@ -9,7 +9,6 @@ namespace ImageProcessorCore
     using System.IO;
 
     using Formats;
-    using Processors;
 
     /// <summary>
     /// Extension methods for the <see cref="Image"/> type.
