+++ src/ImageProcessorCore/Samplers/Processors/IImageSampler.cs
@@ -3,7 +3,7 @@
 // Licensed under the Apache License, Version 2.0.
 // </copyright>
 
-namespace ImageProcessorCore.Processors
+namespace ImageProcessorCore
 {
     /// <summary>
     /// Acts as a marker for generic parameters that require an image sampler.
