+++ src/ImageProcessorCore/Samplers/Processors/RotateProcessor.cs
@@ -1,17 +1,17 @@
-﻿// <copyright file="Skew.cs" company="James Jackson-South">
+﻿// <copyright file="RotateProcessor.cs" company="James Jackson-South">
 // Copyright (c) James Jackson-South and contributors.
 // Licensed under the Apache License, Version 2.0.
 // </copyright>
 
-namespace ImageProcessorCore.Processors
+namespace ImageProcessorCore
 {
     using System.Numerics;
     using System.Threading.Tasks;
 
     /// <summary>
-    /// Provides methods that allow the skewing of images.
+    /// Provides methods that allow the rotating of images.
     /// </summary>
-    public class Skew : ImageSampler
+    public class RotateProcessor : ImageSampler
     {
         /// <summary>
         /// The image used for storing the first pass pixels.
@@ -19,26 +19,21 @@ namespace ImageProcessorCore.Processors
         private Image firstPass;
 
         /// <summary>
-        /// The angle of rotation along the x-axis.
+        /// The angle of rotation in degrees.
         /// </summary>
-        private float angleX;
-
-        /// <summary>
-        /// The angle of rotation along the y-axis.
-        /// </summary>
-        private float angleY;
+        private float angle;
 
         /// <inheritdoc/>
         public override int Parallelism { get; set; } = 1;
 
         /// <summary>
-        /// Gets or sets the angle of rotation along the x-axis in degrees.
+        /// Gets or sets the angle of rotation in degrees.
         /// </summary>
-        public float AngleX
+        public float Angle
         {
             get
             {
-                return this.angleX;
+                return this.angle;
             }
 
             set
@@ -53,33 +48,7 @@ namespace ImageProcessorCore.Processors
                     value += 360;
                 }
 
-                this.angleX = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the angle of rotation along the y-axis in degrees.
-        /// </summary>
-        public float AngleY
-        {
-            get
-            {
-                return this.angleY;
-            }
-
-            set
-            {
-                if (value > 360)
-                {
-                    value -= 360;
-                }
-
-                if (value < 0)
-                {
-                    value += 360;
-                }
-
-                this.angleY = value;
+                this.angle = value;
             }
         }
 
@@ -89,7 +58,7 @@ namespace ImageProcessorCore.Processors
         public Point Center { get; set; }
 
         /// <summary>
-        /// Gets or sets a value indicating whether to expand the canvas to fit the skewed image.
+        /// Gets or sets a value indicating whether to expand the canvas to fit the rotated image.
         /// </summary>
         public bool Expand { get; set; }
 
@@ -102,8 +71,8 @@ namespace ImageProcessorCore.Processors
             {
                 // First find out how big the target rectangle should be.
                 Point centre = this.Center == Point.Empty ? Rectangle.Center(sourceRectangle) : this.Center;
-                Matrix3x2 skew = Point.CreateSkew(centre, -this.angleX, -this.angleY);
-                Rectangle rectangle = ImageMaths.GetBoundingRectangle(sourceRectangle, skew);
+                Matrix3x2 rotation = Point.CreateRotation(centre, -this.angle);
+                Rectangle rectangle = ImageMaths.GetBoundingRectangle(sourceRectangle, rotation);
                 ResizeOptions options = new ResizeOptions
                 {
                     Size = new Size(rectangle.Width, rectangle.Height),
@@ -114,7 +83,7 @@ namespace ImageProcessorCore.Processors
                 Rectangle bounds = ResizeHelper.CalculateTargetLocationAndBounds(source, options);
                 this.firstPass = new Image(rectangle.Width, rectangle.Height);
                 target.SetPixels(rectangle.Width, rectangle.Height, new float[rectangle.Width * rectangle.Height * 4]);
-                new Resize(new NearestNeighborResampler()).Apply(this.firstPass, source, rectangle.Width, rectangle.Height, bounds, sourceRectangle);
+                new ResizeProcessor(new NearestNeighborResampler()).Apply(this.firstPass, source, rectangle.Width, rectangle.Height, bounds, sourceRectangle);
             }
             else
             {
@@ -131,7 +100,7 @@ namespace ImageProcessorCore.Processors
             int startX = 0;
             int endX = this.firstPass.Width;
             Point centre = this.Center == Point.Empty ? Rectangle.Center(this.firstPass.Bounds) : this.Center;
-            Matrix3x2 skew = Point.CreateSkew(centre, -this.angleX, -this.angleY);
+            Matrix3x2 rotation = Point.CreateRotation(centre, -this.angle);
 
             // Since we are not working in parallel we use full height and width 
             // of the first pass image.
@@ -142,11 +111,11 @@ namespace ImageProcessorCore.Processors
                 {
                     for (int x = startX; x < endX; x++)
                     {
-                        // Skew at the centre point
-                        Point skewed = Point.Skew(new Point(x, y), skew);
-                        if (this.firstPass.Bounds.Contains(skewed.X, skewed.Y))
+                        // Rotate at the centre point
+                        Point rotated = Point.Rotate(new Point(x, y), rotation);
+                        if (this.firstPass.Bounds.Contains(rotated.X, rotated.Y))
                         {
-                            target[x, y] = this.firstPass[skewed.X, skewed.Y];
+                            target[x, y] = this.firstPass[rotated.X, rotated.Y];
                         }
                     }
